energy/force consistency check, precision control in trainer
